Maintain trailing slash for non-existent respositories
We add a / to the canonical path but also if we can't find a canonical
path. So we need to be careful that there is not one there already.

This fixes the problem where a trailing / appears in the [respository]
section of the config file, and the dir shows up blank in the dirmodel.