Allow Filejpeg to have multiple pages/files
A JPEG file of the form "base_p<page>.jpeg" will be consider to be a page
of a larger stack of JPEG files. We identify these when opening the desktop,
grouping all matching JPEGs into stacks through an extremely inefficient
n^2 algorithm which will suffice for small directories.

So far, such files do not stack/unstack correct, and there are many other
problems to resolve / features to implement.