PageWidget: Update preview with current page on minor change
The current code will always revert to showing page 1 in the preview when
a stack changed, even when the change is only minor (such as changing the
currently-visible page). Instead, show the current change.

Before this, when you used the delegate page arrows to change from p2 to
p3, the preview would display p1 then p3, a noticeable glitch.