utils: Avoid overwriting the end of a buffer with jpeg_decode()
Odd that this bug has existed for so long, but it is partly due to the fact
that with modern scanners we get an entire JPEG from the file, rather than
the tiled JPEGs we create ourselves.

When a JPEG image height is not a multiple of 16. the JPEG format requires
that it be encoded expanded to the nearest multiple. When decoding we must
ignore anything beyond the end of the image, otherwise we will write beyond
the end of the buffer.

Fix this.