Fix double-uninit of scanner with 'Select scanner'
At present if you click 'Select scanner' in the scan dialog, but then
cancel without selecting a scanner, it deletes the scanner but does not
tell QScannerSetupDlg about it. Then later when QScannerSetupDlg tries
to delete it (e.g. when the dialog is closed or you click
'Select scanner' again) it does a double free, causing a crash in sane.

Fix this by using QScannerSetupDlg to delete the scanner.