qscanner: Update checkButtons() to return an error
When the scanner has been sitting idle for a while, it might turn off.
When that happens SANE cannot communicate with it and saneWordValue()
return INT_MIN to indicate this.

But INT_MIS has all the bits set, so checkButtons() thinks that the
scan button has been pressed, so tries to start a scan. Of course this
fails and an infinite loop arises.

Fix this by propagating the error value up to MainWidget::checkButtons()
and only initiating a scan if there is no error.

Signed-off-by: Simon Glass <sjg@chromium.org>