pscan: Handle a new query while processing an existing one
Since Operation calls processEvents(), if the user types quickly another
search may arrive while the first one is ongoing. Take account of this
and process the search when we can.

This avoids a crash where the Operation is destroyed by one query while
it is still in use by another.

Signed-off-by: Simon Glass <sjg@chromium.org>