pscan: Make keypresses in folder list show up in folderName
Even when typing into the folderName field, the folder list field must
have the focus, since the user will want to scroll up and down, etc.

Pass events on so that this can happen.

Another way to do this would be to have the focus stay in folderName,
with events passed to the folder list. Let's see how it goes.

Signed-off-by: Simon Glass <sjg@chromium.org>