pscan: Add a timer to deal with the folder list
We want to show the folder list in various situations and hide it in
others. Add a timer to check those conditions, so we don't need to add
lots of confusing event handlers in the class.

Show the folder list if the user moves into the folderName field and
there is a valid list.

Hide the folder list if the user moves out of the two relevant fields.

Add a _folders_valid flag to indicate when _folders has valid data from
a search.

Signed-off-by: Simon Glass <sjg@chromium.org>