utils: Add a function to detect date-based matches
We want to present the user with likely patches for a search-string
containing years (4 digits) and months (in the form 03mar). The best
matches are those where both the current year and month are found in the
folder name.

Months are not detected unless a year is present.

Add a function which finds matches, putting the best ones first.

Signed-off-by: Simon Glass <sjg@chromium.org>