pscan: Separate dir creation from scanning
At present mainwidget creates the directory, then does a scaninto it.
This is not ideal, since the _folders list becomes out of date as soon
as the new dir is created.

Split these two functions, with a separate newDir() method in
mainwidget.

Move creation of a missing dir into a function, to keep things tidy.

Drop the unused scanIntoNewDir() function.

Signed-off-by: Simon Glass <sjg@chromium.org>