toolbar: Update to add a butter bar for folder-searching
The feature for filtering by folder name has a reset button, but it does
something quite different to folder searching. It seem better to add a
separate button to reset after a folder search.

Add an extra piece to the bottom of the toolbar, containing this button.
It can be shown as needed.

Signed-off-by: Simon Glass <sjg@chromium.org>