desk: Add a way to refresh a desk
Sometimes files can change without Paperman noticing. For example,
adding a new file in an existing directory does not automatically
infom Paperman of the change.

Add a function to take a another look at a directory. This reads the
paperman file again and then adds any files not mentioned in it.

Signed-off-by: Simon Glass <sjg@chromium.org>