utils: Add a way to scan directories into a cache
When the paper directory is located in an NFS share it can be slow to
access. Add a cache which can remember the location of files and
directories.

For now the test just checks the number of children. Further tests will
be added when the trees can be read and written.

Signed-off-by: Simon Glass <sjg@chromium.org>