utils: Allow reading the file/dir cache from a text file
Add a way to load the contents of the cache from a text file so that it
can be used. This avoids needing to scan the paper directory every time
paperman is started.

For now there is no way to clear the cache.

Expand the tests to cover common cases.

Signed-off-by: Simon Glass <sjg@chromium.org>