dirmodel: Add functions to create a Diritem index
Each Diritem is its own independent model at present. We want the
Dirmodel to be more cohesive, with the model being the same Dirmodel
for all indexes which are passed out.

As a first step towards this, create functions to obtain the root
index for a Diritem, using the Dirmodel to do so.

Since the ordering of items can potentially change, make the root index
emphemeral, calculated just from the row number of the Diritem.

Signed-off-by: Simon Glass <sjg@chromium.org>