dirmodel: Allow storing indexes which are handed out
At present we generate indexes which can be traced to the Diritem which
created them, since the model pointer in the index refers to that
Diritem.

We want to use a single model pointer (the Dirmodel) for every index
which is handed out, so that trick will not work anymore.

Instead, add a map which can turn a Dirmodel index into the associated
Diritem and QDirModel index. Eventually we will add all Dirmodel indexes
to this map.

Signed-off-by: Simon Glass <sjg@chromium.org>